// examples/demo-eos-seed-app/src/generated/schema.ts
export interface ServiceAccountType {
    _: 'serviceAccountType'
    service: string
}
// ...
export interface SubscriptionsCommonInfo {
    _: 'subscriptions.commonInfo'
    am_i_follower: boolean
    is_my_follower: boolean
    followers_count: number
    followings_count: number
}
// ...
export interface UserEmpty {
    _: 'userEmpty'
    id: number
}
// ...
export interface UserService {
    _: 'userService'
    id: number
    self?: true
    contact?: true
    type: ServiceAccountType
    first_name?: string
    last_name?: string
    username?: string
    email: string
    eos_name: string
    photo: UserProfilePhoto
    access_hash: bigint
}
// ...
export interface UserUnreged {
    _: 'userUnreged'
    id: number
    contact?: true
    first_name?: string
    last_name?: string
    phone?: string
    access_hash: bigint
}

export interface User_ {
    _: 'user'
    id: number
    self?: true
    contact?: true
    mutual_contact?: true
    blocked?: true
    verified?: true
    buisness?: true
    first_name?: string
    last_name?: string
    username?: string
    phone?: string
    email: string
    eos_name: string
    photo: UserProfilePhoto
    subscriptions?: SubscriptionsCommonInfo
    lang_code?: string
    access_hash: bigint
    kyc_verifed: number
    has_unionpay_account?: true
    status?: UserStatus
}
// ...
export type User = UserEmpty | UserService | UserUnreged | User_

export type UserProfilePhoto = UserProfilePhotoEmpty | UserProfilePhoto_

export type UserStatus =
    | UserStatusEmpty
    | UserStatusOnline
    | UserStatusOffline
    | UserStatusRecently
    | UserStatusLastWeek
    | UserStatusLastMonth
// ...
export interface CryptoSendCodeParams {
    public_key: string
    api_id: number
    api_hash: string
}

export interface CryptoSignInParams {
    public_key: string
    code: string
    server_sign: string
    sign: string
}

export interface CryptoSignUpParams {
    public_key: string
    code: string
    server_sign: string
    sign: string
    phone_number: string
    first_name: string
    last_name: string
    email: string
    username: string
    country?: string
}